Moving test utils into _test file, to align with lint requirements
